Fixes #60 refactors the heck out of transform-ast.js
This refactors transform-ast.js into many many files within the src/transform/ folder. It also has the unrelated change of making 'require' its own keyword so it will compile differently and Browserify will work correctly. May have changed other stuff too honestly because dang there was just a lot going on here. I also got jshint properly configured so that allowed me to fix some outstanding code issues I hadn't noticed yet
