+++ examples/bst.squiggle
@@ -10,13 +10,15 @@ let Tip = {type: "Tip"}
 def leaf(x) =
     Node(x, Tip, Tip)
 
-def bstAdd(d, n) =
+# This actually doesn't end up looking much better with pattern matching since
+# we need to make the "less-than" check on the data.
+def bstAdd(n, d) =
     if n == Tip then
         leaf(d)
     else if d < n.data then
-        Node(n.data, bstAdd(d, n.left), n.right)
+        Node(n.data, bstAdd(n.left, d), n.right)
     else
-        Node(n.data, n.left, bstAdd(d, n.right))
+        Node(n.data, n.left, bstAdd(n.right, d))
 
 def inOrder(node) =
     match node
@@ -25,9 +27,9 @@ def inOrder(node) =
         inOrder(left) ++ [data] ++ inOrder(right)
     end
 
-let root = foldRight(bstAdd, Tip, [4, 2, 1, 3, 6, 5, 7])
+def treeFolder(x, node) =
+    bstAdd(node, x)
 
-def _add(x, y) = x + y
-def _mul(x, y) = x * y
+let root = foldRight([4, 2, 1, 3, 6, 5, 7], Tip, treeFolder)
 
 in console.log(inOrder(root))
