+++ src/js-reserved-words.json
@@ -10,6 +10,7 @@
     "true",
     "false",
     "return",
+    "require",
     "if",
     "else",
     "enum",
