+++ src/transform/identifier.js
@@ -1,0 +1,15 @@
+var es = require("../es");
+var isJsReservedWord = require("../is-js-reserved-word");
+
+function cleanIdentifier(s) {
+    if (isJsReservedWord(s)) {
+        return s + "_";
+    }
+    return s;
+}
+
+function Identifier(transform, node) {
+    return es.Identifier(cleanIdentifier(node.data));
+}
+
+module.exports = Identifier;
