+++ src/transform/negate.js
@@ -1,0 +1,9 @@
+var es = require("../es");
+
+function Negate(transform, node) {
+    var expr = transform(node.expr);
+    var negate = es.Identifier("$negate");
+    return es.CallExpression(negate, [expr]);
+}
+
+module.exports = Negate;
