+++ src/transform/null.js
@@ -1,0 +1,9 @@
+var es = require("../es");
+
+var nil = es.Literal(null);
+
+function Null(transform, node) {
+    return nil;
+}
+
+module.exports = Null;
