+++ src/transform/number.js
@@ -1,0 +1,7 @@
+var es = require("../es");
+
+function Number_(transform, node) {
+    return es.Literal(node.data);
+}
+
+module.exports = Number_;
