+++ src/transform/parameter.js
@@ -1,0 +1,5 @@
+function Parameter(transform, node) {
+    return transform(node.identifier);
+}
+
+module.exports = Parameter;
