+++ src/transform/repl-expression.js
@@ -1,0 +1,7 @@
+var fileWrapper = require("../file-wrapper");
+
+function ReplExpression(transform, node) {
+    return fileWrapper([transform(node.expression)]);
+}
+
+module.exports = ReplExpression;
