+++ src/transform/string.js
@@ -1,0 +1,7 @@
+var es = require("../es");
+
+function String_(transform, node) {
+    return es.Literal(node.data);
+}
+
+module.exports = String_;
