+++ src/transform/true.js
@@ -1,0 +1,9 @@
+var es = require("../es");
+
+var theTruth = es.Literal(true);
+
+function True(transform, node) {
+    return theTruth;
+}
+
+module.exports = True;
