+++ src/transform/undefined.js
@@ -1,0 +1,9 @@
+var es = require("../es");
+
+var nothing = es.Identifier("undefined");
+
+function Undefined(transform, node) {
+    return nothing;
+}
+
+module.exports = Undefined;
